+++ gh-pages/js/examples-renderer.js
@@ -52532,8 +52532,8 @@ function fun (graph, additionalFunctions) {
   injectDotOperators(funcs, task)
   injectGlobals(funcs, task)
   injectReferences(funcs, task)
-  injectNumbers(funcs, task)
   injectStrings(funcs, task)
+  injectNumbers(funcs, task)
   injectArrowFunctions(funcs, task)
 
   /**
@@ -53109,6 +53109,8 @@ function injectDotOperators (funcs, task) {
      */
 
     function dotOperatorAttributeWrite (attributeName, obj, attributeValue) {
+      if (arguments.length === 1) return
+
       obj[attributeName] = attributeValue
 
       return obj
@@ -53679,6 +53681,8 @@ module.exports = validate
 
 var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
 
+var regexQuoted = require('./regex/quoted');
+
 var globalContext;
 
 if ((typeof window === 'undefined' ? 'undefined' : _typeof(window)) === 'object') {
@@ -53699,8 +53703,12 @@ if ((typeof global === 'undefined' ? 'undefined' : _typeof(global)) === 'object'
  */
 
 function walkGlobal(taskName) {
-  // Skip dot operator and tasks that starts with a dot.
+  // Skip dot operator and tasks that start with a dot.
   if (taskName.indexOf('.') === 0) return;
+
+  // Skip strings and numbers which may include dots.
+  if (regexQuoted.test(taskName)) return;
+  if (parseFloat(taskName)) return;
 
   function toNextProp(next, prop) {
     return next[prop];
@@ -53712,7 +53720,7 @@ function walkGlobal(taskName) {
 module.exports = walkGlobal;
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{}],320:[function(require,module,exports){
+},{"./regex/quoted":314}],320:[function(require,module,exports){
 module.exports={
   "data": {
     "results": [ {