+++ gh-pages/changelog.md
@@ -28,6 +28,7 @@ Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 ### Fixed
 
 - Graph.update() now is async, it caused a test to fail randomly
+- Quoted tasks can contain dots, there was a conflict with global tasks
 
 ### Added